Rename parser/parser.py -> parser/parser_module.py
