+++ python/python/opm/io/parser/__init__.py
@@ -1,4 +1,4 @@
 from opm._common import action
 
-from .parser import parse, load_deck, load_deck_string, parse_string
+from .parser_module import parse, load_deck, load_deck_string, parse_string
